파일 업로드 기능 추가 및 에러관리

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
   const { search } = req.query;
@@ -17,7 +18,8 @@ export const home = async (req, res) => {
 
 export const getWatch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("uploader");
+  console.log(video);
   if (!video) {
     return res
       .status(404)
@@ -32,7 +34,11 @@ export const getUpload = (req, res) => {
 
 export const postUpload = async (req, res) => {
   const { title, description, hashtags } = req.body;
-  await Video.create({
+  const uploader = req.session.user._id;
+  const { path } = req.file;
+  const newVideo = await Video.create({
+    uploader,
+    videoUrl: path,
     title,
     description,
     createdAt: Date.now(),
@@ -42,12 +48,22 @@ export const postUpload = async (req, res) => {
       rating: 0,
     },
   });
+  const user = await User.findById(uploader);
+  user.videos.push(newVideo._id);
+  user.save();
   return res.status(201).redirect("/");
 };
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  if (!req.session.user) {
+    return res.status(403).redirect("/");
+  }
+  const { _id } = req.session.user;
   const video = await Video.findById(id);
+  if (String(video.uploader) !== _id) {
+    return res.status(403).redirect("/");
+  }
   return res.render("videos/edit-video", { pageTitle: "Edit Video", video });
 };
 
@@ -64,6 +80,14 @@ export const postEdit = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  if (!req.session.user) {
+    return res.status(403).redirect("/");
+  }
+  const { _id } = req.session.user;
+  const video = await Video.findById(id);
+  if (String(video.uploader) !== _id) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.status(200).redirect("/");
 };
